Implementation of Spring security end-to-end / scripts handling auth and data on frontend
The code involves incorporating spring security library into the project. It has major impact on entire project from end to end. The scripts were include to handle the data received from the API and include token on requests.

There is no session, its a stateless communication between frontend and backend. The token is stored in a cookie which frontend has to be sure is included which each call.

The commit was tested end-to-end, from creation of new account to login.
Fixed some issues on welcome and registration pages to make the end-to-end work.

Its built on top of latest changes from Dara