Handle currency conversion on frontend/backend
New feature:
New endpoint that let users make a request for currency exchange. It will update both accounts balance and create 2 transactions.
The frontend has a new button with a modal that handles the conversion. The rate is calculated in the modal too for the user to see the resulting conversion directly, before submitting the request.
The rates are calculated on the backend to avoid issues as a precaution.